support waitUntil option for `page.setContent`
