+++ lib/PuppeteerSharp.Tests/PageTests/SetContentTests.cs
@@ -44,5 +44,20 @@ namespace PuppeteerSharp.Tests.PageTests
 
             Assert.Equal($"{doctype}{ExpectedOutput}", result);
         }
+
+        [Fact]
+        public async Task ShouldAwaitResourceToLoad()
+        {
+            var imgPath = "/img.png";
+            var imgResponse = new TaskCompletionSource<bool>();
+            Server.SetRoute(imgPath, context => imgResponse.Task);
+            var loaded = false;
+            var contentTask = Page.SetContentAsync($"<img src=\"${TestConstants.ServerUrl + imgPath}\"></img>")
+                .ContinueWith(_ => loaded = true);
+            await Server.WaitForRequest(imgPath);
+            Assert.False(loaded);
+            imgResponse.SetResult(true);
+            await contentTask;
+        }
     }
 }
