+++ lib/PuppeteerSharp/Page.cs
@@ -711,9 +711,10 @@ namespace PuppeteerSharp
         /// Sets the HTML markup to the page
         /// </summary>
         /// <param name="html">HTML markup to assign to the page.</param>
+        /// <param name="options">The navigations options</param>
         /// <returns>Task.</returns>
-        /// <seealso cref="Frame.SetContentAsync(string)"/>
-        public Task SetContentAsync(string html) => _frameManager.MainFrame.SetContentAsync(html);
+        /// <seealso cref="Frame.SetContentAsync(string, SetContentOptions)"/>
+        public Task SetContentAsync(string html, SetContentOptions options = null) => _frameManager.MainFrame.SetContentAsync(html);
 
         /// <summary>
         /// Navigates to an url
